requirements.txt
python-whois
requests
dnslib
tldextract
dnstwist
dnspython
bs4